Added caching of processed GitHub API responses
In the helper component tried to set up the caching framework for
responses from the GitHub API.  Should be flexible enough to add other
feature in the future besides just issues if desired.  Put in code to do
conditional requests and a minimum interval between requests for the
same information.  The latter period is controlled by a configuration
variable, defaulting to 10 minutes.

Still need to add the actual parsing of the JSON response into the
display table.  Also handling pagination of the responses.

Also handling non-200 or 304 responses is basically through up your
hands and give up, so that should be tidied.